Move info pane content to subfolder. Animate on open.
