+++ app/assets/javascripts/tiptap/InfoPane/EmailLinks.js
