+++ app/assets/javascripts/tiptap/editor.css
@@ -1,4 +1,7 @@
 /* Basic editor styles */
+:root {
+  interpolate-size: allow-keywords;
+}
 .tiptap {
   min-height: 400px;
   &:focus {
@@ -355,6 +358,15 @@
 }
 
 .info-pane {
+  height: auto;
+  overflow: hidden;
+  transition: height 0.4s ease-out;
+
+  /* 3. The initial state of the element upon insertion into the DOM */
+  @starting-style {
+    height: 0;
+  }
+  
   @apply bg-gray-200;
 
   p {
